+++ internal/driver/mobile/device.go
@@ -41,10 +41,6 @@ func (*device) HasKeyboard() bool {
 	return false
 }
 
-func (*device) SetDisableScreenBlanking(disable bool) {
-	setDisableScreenBlank(disable)
-}
-
 func (*device) ShowVirtualKeyboard() {
 	showVirtualKeyboard(mobile.DefaultKeyboard)
 }
