+++ internal/driver/mobile/device_android.go
@@ -2,18 +2,7 @@
 
 package mobile
 
-import (
-	"fyne.io/fyne/v2"
-	"fyne.io/fyne/v2/driver"
-)
-
-/*
-#include <stdbool.h>
-#include <stdlib.h>
-
-void keepScreenOn(uintptr_t jni_env, uintptr_t ctx, bool disabled);
-*/
-import "C"
+import "fyne.io/fyne/v2"
 
 const tapYOffset = -12.0 // to compensate for how we hold our fingers on the device
 
@@ -28,14 +17,4 @@ func (*device) SystemScaleForWindow(_ fyne.Window) float32 {
 		return 1.5
 	}
 	return 1
-}
-
-func setDisableScreenBlank(disable bool) {
-	driver.RunNative(func(ctx any) error {
-		ac := ctx.(*driver.AndroidContext)
-
-		C.keepScreenOn(C.uintptr_t(ac.Env), C.uintptr_t(ac.Ctx), C.bool(disable))
-
-		return nil
-	})
 }
