+++ internal/driver/mobile/driver_ios.m
@@ -1,0 +1,12 @@
+//go:build ios
+
+#import <Foundation/Foundation.h>
+#import <UIKit/UIKit.h>
+
+void disableIdleTimer(BOOL disabled) {
+    @autoreleasepool {
+        [[NSOperationQueue mainQueue] addOperationWithBlock:^ {
+            [UIApplication sharedApplication].idleTimerDisabled = disabled;
+        }];
+    }
+}
